- 폴더 시작 진입점 도메인 기준으로 변경 - 헥사고날 의미 수준으로 도메인과 도메인 엔티티 분리 ( 매핑을 위한 dto, entity Mapper 추가)
