--- src/main/java/com/nettee/adapter/in/web/board/dto/BoardCreateCommand.java
@@ -1,9 +1,0 @@
-package com.nettee.adapter.in.web.board.dto;
-
-import jakarta.validation.constraints.NotNull;
-
-public record BoardCreateCommand(
-        @NotNull String title,
-        @NotNull String content
-) {
-}
