--- src/main/java/com/nettee/adapter/in/web/board/dto/BoardDeleteCommand.java
@@ -1,8 +1,0 @@
-package com.nettee.adapter.in.web.board.dto;
-
-import jakarta.validation.constraints.NotNull;
-
-public record BoardDeleteCommand(
-        @NotNull Long id
-) {
-}
