--- src/main/java/com/nettee/adapter/in/web/board/dto/BoardDto.java
@@ -1,8 +1,0 @@
-package com.nettee.adapter.in.web.board.dto;
-
-public record BoardDto (
-        Long id,
-        String title,
-        String content
-) {
-}
