--- src/main/java/com/nettee/application/port/BoardCommandPort.java
@@ -1,12 +1,0 @@
-package com.nettee.application.port;
-
-import com.nettee.application.domain.board.Board;
-
-public interface BoardCommandPort {
-
-    Board createBoard(Board board);
-
-    Board updateBoard(Board board);
-
-    void deleteBoard(Long id);
-}
