--- src/main/java/com/nettee/application/usecase/BoardCreateUseCase.java
@@ -1,8 +1,0 @@
-package com.nettee.application.usecase;
-
-import com.nettee.application.domain.board.Board;
-
-public interface BoardCreateUseCase {
-
-    Board createBoard (Board board);
-}
