--- src/main/java/com/nettee/application/usecase/BoardDeleteUseCase.java
@@ -1,6 +1,0 @@
-package com.nettee.application.usecase;
-
-public interface BoardDeleteUseCase {
-
-    void deleteBoard (Long boardId);
-}
