--- src/main/java/com/nettee/application/usecase/BoardUpdateUseCase.java
@@ -1,9 +1,0 @@
-package com.nettee.application.usecase;
-
-import com.nettee.application.domain.board.Board;
-
-public interface BoardUpdateUseCase {
-
-    Board updateBoard (Board board);
-
-}
