+++ src/main/java/com/nettee/board/adapter/in/web/dto/BoardCreateCommand.java
@@ -1,0 +1,9 @@
+package com.nettee.board.adapter.in.web.dto;
+
+import jakarta.validation.constraints.NotNull;
+
+public record BoardCreateCommand(
+        @NotNull String title,
+        @NotNull String content
+) {
+}
