+++ src/main/java/com/nettee/board/adapter/in/web/dto/BoardDeleteCommand.java
@@ -1,0 +1,8 @@
+package com.nettee.board.adapter.in.web.dto;
+
+import jakarta.validation.constraints.NotNull;
+
+public record BoardDeleteCommand(
+        @NotNull Long id
+) {
+}
