+++ src/main/java/com/nettee/board/adapter/in/web/dto/BoardDto.java
@@ -1,0 +1,8 @@
+package com.nettee.board.adapter.in.web.dto;
+
+public record BoardDto (
+        Long id,
+        String title,
+        String content
+) {
+}
