+++ src/main/java/com/nettee/board/application/usecase/BoardCreateUseCase.java
@@ -1,0 +1,8 @@
+package com.nettee.board.application.usecase;
+
+import com.nettee.board.application.domain.Board;
+
+public interface BoardCreateUseCase {
+
+    Board createBoard (Board board);
+}
