+++ src/main/java/com/nettee/board/application/usecase/BoardDeleteUseCase.java
@@ -1,0 +1,6 @@
+package com.nettee.board.application.usecase;
+
+public interface BoardDeleteUseCase {
+
+    void deleteBoard (Long boardId);
+}
