+++ src/main/java/com/nettee/board/application/usecase/BoardUpdateUseCase.java
@@ -1,0 +1,9 @@
+package com.nettee.board.application.usecase;
+
+import com.nettee.board.application.domain.Board;
+
+public interface BoardUpdateUseCase {
+
+    Board updateBoard (Board board);
+
+}
